requirements.txt
requests
websocket-client
pandas
numpy
tensorflow
scikit-learn
python-binance
python-dotenv